requirements.txt
# ── Python packages ───────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.40.0
ultralytics>=8.3.0
opencv-python-headless>=4.9.0
